# speed up `make_replicate_tbl()`